requirements.txt
email-validator>=2.3.0
flask>=3.1.2
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.11
werkzeug>=3.1.4
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email-validator
flask
flask-login
flask-sqlalchemy
gunicorn
psycopg2-binary
werkzeug
python-dotenv
flask-wtf
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email-validator
flask
flask-login
flask-sqlalchemy
gunicorn
psycopg2-binary
werkzeug
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email-validator
flask
flask-login
flask-sqlalchemy
gunicorn
psycopg2-binary
werkzeug
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email-validator
flask
flask-login
flask-sqlalchemy
gunicorn
psycopg2-binary
werkzeug
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email-validator
flask
flask-login
flask-sqlalchemy
gunicorn
psycopg2-binary
werkzeug
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email-validator
flask
flask-login
flask-sqlalchemy
gunicorn
psycopg2-binary
werkzeug
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email-validator
flask
flask-login
flask-sqlalchemy
flask-wtf
gunicorn
psycopg2-binary
python-dotenv
werkzeug
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email-validator
flask
flask-login
flask-sqlalchemy
gunicorn
psycopg2-binary
werkzeug
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email-validator
flask
flask-login
flask-sqlalchemy
gunicorn
psycopg2-binary
werkzeug
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email-validator
flask
flask-login
flask-sqlalchemy
gunicorn
psycopg2-binary
werkzeug
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
flask-limiter
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email-validator
flask
flask-login
flask-sqlalchemy
gunicorn
psycopg2-binary
werkzeug
